+++ include/nstd/core/cstr/cstr.h
@@ -528,6 +528,10 @@ NSTDAPI const NSTDChar *nstd_core_cstr_mut_get_null_const(const NSTDCStr *cstr);
 
 /// Return a pointer the character at `pos` in `cstr`.
 ///
+/// # Note
+///
+/// This will return a null pointer if `pos` is greater than `NSTDInt`'s max value.
+///
 /// # Parameters:
 ///
 /// - `NSTDCStrMut *cstr` - The C string.