+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 - Added `math_pow_*`.
 - Added `math_abs_*`.
 ### `nstd.os`
-- Added `NSTD_OS_[LINUX|WINDOWS]` constants.
+- Added `NSTD_OS_[LINUX|MACOS|WINDOWS]` constants.
 
 # 0.1.3
 ### `nstd`